Pull more functions over to the facade
